Initiliazed the singin screen
